refactor: 原版对齐批次1 — 261 个文件
用户指令: 全仓对齐 RikkaHub 原版, 贯彻 可改可不改必须改;
有意义的优化保留, 没意义的对齐回根。

批次1 范围:
- 255 个 code=0 文件 (仅注释/空行差异) 全量回原版 — 清理
  对齐标注批注与多余空白, 行为零变化
- 6 个等价改写文件回原版: ErrorParser (else 冗余) /
  AIRequestInterceptor (行合并) / McpOAuthClient (?.string 防御) /
  MigrationUtils (智能转换注释) / ViewText (toFloat) /
  WorkspaceTerminalSession (SHOW_IMPLICIT)

保留 (有意义优化, 计入对齐清单):
- SearchTools 日期注入删除 (v3.3.2 缓存优化)
- CalendarTool needsApproval=false / OcrTransformer 兼容 /
  MemoryEntity Long id / ConversationDAO 增量 / CardGroup 编译适配等 19 项

后续批次: REVIEW 清单 78 文件逐个人工判定 (RouteActivity/McpManager/
ChatService/UI 页面等大 diff 文件分批执行)

+++ app/src/main/java/me/rerere/rikkahub/ui/components/ui/ViewText.kt
@@ -1,9 +1,5 @@
 package me.rerere.rikkahub.ui.components.ui
 
-
-/* ───【原版对齐】ViewText.kt | 差异 ±2 行
- * 来源: 原版移植 + 自研小调整 (未达专项标注阈值, 对齐细节见对齐地图)
- * ───────────────────────────────────────────────────────────────*/
 import android.graphics.Paint
 import android.graphics.Typeface
 import android.graphics.text.LineBreaker
@@ -191,7 +187,7 @@ private fun TextView.setComposeTextStyle(
                 // extra = desired_line_height - font_metrics_height
                 val fontMetrics = paint.fontMetricsInt
                 val extraSpacing = lineHeightPx - (fontMetrics.descent - fontMetrics.ascent)
-                setLineSpacing(extraSpacing, 1.0f)
+                setLineSpacing(extraSpacing.toFloat(), 1.0f)
             }
         }
